Client Order(GOA-70), interface has been made, curently no database
