--- Users/C-123/password.txt
@@ -1,1 +1,0 @@
-asd
