--- Users/C-ad/password.txt
@@ -1,1 +1,0 @@
-b
