--- Users/C-adsg/password.txt
@@ -1,1 +1,0 @@
-das
