--- Users/M-ads/password.txt
@@ -1,1 +1,0 @@
-fds
