--- Users/M-manag/password.txt
@@ -1,1 +1,0 @@
-1234
