+++ src/main/java/Main.java
@@ -1,9 +1,10 @@
 
-
+import startScreen.StartScreen;
 
 public class Main {
     public static void main(String[] args)
     {
-        System.out.println("ceva");
+        System.out.println( com.sun.javafx.runtime.VersionInfo.getRuntimeVersion());
+
     }
 }
