+++ src/main/java/register/MakeAccount.java
@@ -1,8 +1,7 @@
 package register;
 
 import java.io.*;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+
 
 public class MakeAccount {
