+++ src/main/java/register/RegisterInterface.java
@@ -64,6 +64,7 @@ public class RegisterInterface {
             if(InputCheck.badImput(nameImput.getText(),pasImput.getText())) return;
 
 
+            System.out.println("TEST");
             boolean check;
             if(typeImput.getValue().equals("Client")) check = true;
             else check = false;
